requirements.txt
streamlit
google-generativeai
pypdf
graphviz